# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "iKon Mall Backend",
        "description": "iKon Mall Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "locahost:5000/api",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/brand/{brandId}": {
            "delete": {
                "tags": [
                    "Brands"
                ],
                "summary": "Delete Brand by Id",
                "description": "",
                "operationId": "deleteBrand",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "brandId",
                        "in": "path",
                        "description": "ID of the address that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Brands"
                ],
                "summary": "Updates brand",
                "description": "",
                "operationId": "updatebrand",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "brandId",
                        "in": "path",
                        "description": "brandId that need to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated brand object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Brand"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/category": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "adds a new category",
                "description": "",
                "operationId": "addCategory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Category object that needs to be added",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Error"
                    }
                }
            },
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "get all categories",
                "description": "",
                "operationId": "getCategory",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "skip",
                        "type": "integer",
                        "description": "The number of documents to skip before starting to collect the result set"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "type": "integer",
                        "description": "The numbers of documents to return"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Error in getting Category"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Brand": {
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"                    
                },
                "description": {
                    "type": "string"
                },
                "richDescription": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                }
            }
        },
...
        "Category": {
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Electronics"
                },
                "description": {
                    "type": "string",
                    "example": "A category for electronics items"
                },
                "richDescription": {
                    "type": "string"
                }
            }
        },
...
    }
}